feat: standarize project after a lot of changes, introduce new view
