+++ src/murawa/services/rendering/__init__.py
@@ -1,0 +1,15 @@
+from murawa.services.rendering.overlay import (
+    build_tracking_marker_label,
+    team_overlay_summary,
+    write_annotated_match_video,
+    write_preview_assets,
+    write_team_assignment_debug_preview,
+)
+
+__all__ = [
+    "build_tracking_marker_label",
+    "team_overlay_summary",
+    "write_annotated_match_video",
+    "write_preview_assets",
+    "write_team_assignment_debug_preview",
+]
