+++ src/murawa/services/runtime/saved_match_analyses.py
@@ -5,39 +5,56 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any
 
+from murawa.settings import OUTPUTS_PREDICTIONS, OUTPUTS_VIDEOS
+
 MATCH_VIDEO_SUFFIXES = {".mp4", ".webm"}
 
 
 @dataclass(frozen=True)
 class SavedMatchAnalysis:
     analysis_id: str
-    video_path: Path
+    preview_path: Path
+    download_path: Path | None
     modified_at_ns: int
     summary_path: Path | None
     summary: dict[str, Any] | None
 
+    @property
+    def video_path(self) -> Path:
+        """Backward-compatible alias for browser preview path."""
+        return self.preview_path
+
 
 def list_saved_match_analyses(project_root: Path) -> list[SavedMatchAnalysis]:
-    videos_dir = project_root / "outputs" / "videos"
-    predictions_dir = project_root / "outputs" / "predictions"
+    videos_dir = project_root / OUTPUTS_VIDEOS
+    predictions_dir = project_root / OUTPUTS_PREDICTIONS
     if not videos_dir.exists() or not videos_dir.is_dir():
         return []
 
-    saved_analyses: list[SavedMatchAnalysis] = []
+    grouped_videos: dict[str, dict[str, Path]] = {}
     for video_path in videos_dir.iterdir():
         if not _is_match_video(video_path):
             continue
+        grouped_videos.setdefault(video_path.stem, {})[video_path.suffix.lower()] = video_path
 
+    saved_analyses: list[SavedMatchAnalysis] = []
+    for analysis_id, paths in grouped_videos.items():
+        preview_path = paths.get(".webm")
+        if preview_path is None:
+            continue
+
+        download_path = paths.get(".mp4")
         try:
-            modified_at_ns = video_path.stat().st_mtime_ns
+            modified_at_ns = preview_path.stat().st_mtime_ns
         except OSError:
             continue
 
-        summary_path = _find_summary_path(predictions_dir, analysis_id=video_path.stem)
+        summary_path = _find_summary_path(predictions_dir, analysis_id=analysis_id)
         saved_analyses.append(
             SavedMatchAnalysis(
-                analysis_id=video_path.stem,
-                video_path=video_path,
+                analysis_id=analysis_id,
+                preview_path=preview_path,
+                download_path=download_path,
                 modified_at_ns=modified_at_ns,
                 summary_path=summary_path,
                 summary=_load_summary(summary_path),
@@ -46,7 +63,7 @@ def list_saved_match_analyses(project_root: Path) -> list[SavedMatchAnalysis]:
 
     return sorted(
         saved_analyses,
-        key=lambda analysis: (analysis.modified_at_ns, analysis.video_path.name),
+        key=lambda analysis: (analysis.modified_at_ns, analysis.preview_path.name),
         reverse=True,
     )
