+++ src/murawa/services/vision/team_assignment_helpers.py
